# api/models/model.go
package models

type ErrorResponse struct {
	Message string `json:"message"`
	Code    string `json:"code"`
}

type SuccessResponse struct {
	Message string `json:"message"`
}

type ID struct {
	Id string `json:"id"`
}